+++ cfme/classic_ui/__init__.py
@@ -4,45 +4,8 @@ from pomoc.patternfly import Button
 from pomoc.navigator import Navigator
 from pomoc.objects import View
 
-from .widgets import TwoLevelMenuItem
-
-
-class CFMEView(View):
-    class menu(View):
-        dashboard = TwoLevelMenuItem('Cloud Intel', 'Dashboard')
-        reports = TwoLevelMenuItem('Cloud Intel', 'Reports')
-        chargeback = TwoLevelMenuItem('Cloud Intel', 'Chargeback')
-        timelines = TwoLevelMenuItem('Cloud Intel', 'Timelines')
-        rss = TwoLevelMenuItem('Cloud Intel', 'RSS')
-
-        @Navigator.transition_to('Dashboard')
-        def go_to_dashboard(self):
-            self.dashboard.click()
-
-        @Navigator.transition_to('Reports')
-        def go_to_reports(self):
-            self.reports.click()
-
-        # insights = TopMenuItem('Red Hat Insights')
-        # services = TopMenuItem('Services')
-        # compute = TopMenuItem('Compute')
-        # configuration = TopMenuItem('Configuration')
-        # networks = TopMenuItem('Networks')
-        # control = TopMenuItem('Control')
-        # automate = TopMenuItem('Automate')
-        # optimize = TopMenuItem('Optimize')
-        # settings = TopMenuItem('Settings')
-
-
-class Dashboard(CFMEView):
-    reset_button = Button(title='Reset Dashboard Widgets to the defaults')
-
-    def on_view(self):
-        return self.reset_button.is_displayed
-
-
-class Reports(CFMEView):
-    pass
+from .intelligence import Dashboard
+from .compute.infrastructure.providers import InfrastructureProviders
 
 
 class Login(View):
@@ -53,7 +16,7 @@ class Login(View):
 
     # Currently, the Menu will be separate views since this would have to lead to many many
     # possible targets ....
-    @Navigator.transition_to(Dashboard)
+    @Navigator.transition_to(Dashboard, InfrastructureProviders)
     def login_user(self, user):
         self.username.fill(user.credential.principal)
         self.password.fill(user.credential.secret)
