+++ cfme/classic_ui/compute/infrastructure/providers.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+from cfme.classic_ui import CFMEView
+from cfme.classic_ui.widgets import ContentTitle
+
+
+class InfrastructureProviders(CFMEView):
+    title = ContentTitle('Infrastructure Providers')
+
+    def on_view(self):
+        return self.title.is_displayed
